+++ docs/source/command_line_documentation.rst
@@ -1,0 +1,12 @@
+Documentation of the command line interface of Instrumentum sanae Doctrinae 
+===========================================================================
+
+
+Documentation of the monergism command 
+--------------------------------------
+
+The **monergism** command allows the user to interact with `Monergism <https://www.monergism.com>`_.
+There are many positionnal parameters for this command. Here are they 
+
+1
+
