+++ docs/source/index.rst
@@ -18,5 +18,6 @@ documentation for details.
    Instrumentum_sanae_doctrinae.web_scraping
    json_data_structure_documentation
    modules
+   command_line_documentation
