+++ src/Instrumentum_sanae_doctrinae/cli_interface/argument_parser.py
@@ -1,7 +1,7 @@
 import asyncio
 import os
+from Instrumentum_sanae_doctrinae.cli_interface.web_scraping.monergism_command_executer import *
 import click
-from Instrumentum_sanae_doctrinae.web_scraping.monergism import mn_scrap_general_information, mn_scrap_get_list
 from Instrumentum_sanae_doctrinae.web_scraping.my_constants import SCRIPTURE_NAME, SPEAKER_NAME, TOPIC_NAME
 
 
@@ -16,89 +16,34 @@ from Instrumentum_sanae_doctrinae.web_scraping.my_constants import SCRIPTURE_NAM
 """
 
 
-#root_folder = os.path.join(os.getcwd(),'test_folder')
-
-#print(root_folder)
-
-def monergism_scrap_list(root_folder):
-    # Get the list of the topics 
-    ob = mn_scrap_get_list.GetTopicList(root_folder)
-    asyncio.run(ob.scrap_and_write())
-    
-    # Get the list of speakers
-    ob = mn_scrap_get_list.GetSpeakerList(root_folder)
-    asyncio.run(ob.scrap_and_write())
-    
-    # Get the list of the scriptures 
-    ob = mn_scrap_get_list.GetScriptureList(root_folder)
-    asyncio.run(ob.scrap_and_write())
-    
-    
-    
-def monegism_scrap_general_information(root_folder,browse_by_type,
-                                       download_batch_size,
-                                       overwrite_log):
-    """_summary_
-
-    Args:
-        root_folder (_type_): The root folder where the download and metadata, logs data are stored 
-        browse_by_type (_type_): the browse by type (either scripture, topic or speaker)
-        subject (_type_): optionnal. 
-    """
-    print(root_folder,browse_by_type,overwrite_log)
-    if browse_by_type in [SPEAKER_NAME,SCRIPTURE_NAME,TOPIC_NAME]:
-        
-        ob = mn_scrap_general_information.MonergismScrapGeneralInformation_ALL(
-            root_folder = root_folder,
-            browse_by_type = browse_by_type,
-            overwrite_log = overwrite_log
-        )
-        print(ob.__dict__)
-        asyncio.run(ob.print_download_informations(True))
-        asyncio.run(ob.download(download_batch_size=download_batch_size))
-        
-        
-    else:
-        a = 0
-
+# The general group. It is the basis group 
 @click.group()
-def cli():
+def entry_point():
     pass 
 
 
-@cli.command(name = "monergism",help = """Interact with monergism.com""")
-@click.argument("action",required = True)
-@click.argument("browse_by_type",required = False)
-@click.argument("output_folder",required = True)
-@click.option('-u', '--overwrite-log',"overwrite_log",
-              default = False,required =  False)
-@click.option("-bs","--download_batch_size","download_batch_size",
-              default = 10,type = int,required = False)
 
-def configure_monergism_arguments(action,output_folder,browse_by_type,
-                                  overwrite_log,download_batch_size):
+
+
     
-    if "=" in output_folder:
-        output_folder = output_folder.split("=")[1]
-    if action == "scrap_list":
-        a = 0
-        #monergism_scrap_list(output_folder)
-    elif action == "scrap_general_info":
-        if browse_by_type:
-            monegism_scrap_general_information(root_folder = output_folder,
-                                           browse_by_type=browse_by_type,
-                                           download_batch_size = download_batch_size,
-                                           overwrite_log = overwrite_log)
-        else:
-            raise ValueError("The argument <browse_by_type> must be given if the action parameter is <scrap_general_info>")
-        
-    
-   
+ 
+# The group for monergism commands
+@entry_point.group(name = "monergism")
+def monergism_group():
+    pass 
 
 
+# The group for sermonindex commands
+@entry_point.group(name = "sermonindex")
+def sermon_index_group():
+    pass 
 
-        
+
+# Add command for monergism
+monergism_group.add_command(scrap_list_command,name="scrap_list")
+monergism_group.add_command(scrap_general_information_command,name="scrap_general_information")
+
 
 if __name__ == "__main__":
-    cli()
+    entry_point()
