--- src/Instrumentum_sanae_doctrinae/cli_interface/test.py
@@ -1,124 +1,0 @@
-import click
-
-@click.group(context_settings={"help_option_names": ["--help", "-h"]})
-def git():
-    """Git - the stupid content tracker"""
-    pass
-
-# --- Start a Working Area ---
-@git.command()
-def clone():
-    """Clone a repository into a new directory"""
-    click.echo("Cloning repository...")
-
-@git.command()
-def init():
-    """Create an empty Git repository or reinitialize an existing one"""
-    click.echo("Initializing repository...")
-
-# --- Work on the Current Change ---
-@git.command()
-def add():
-    """Add file contents to the index"""
-    click.echo("Adding files to index...")
-
-@git.command()
-def mv():
-    """Move or rename a file, a directory, or a symlink"""
-    click.echo("Moving/Renaming...")
-
-@git.command()
-def restore():
-    """Restore working tree files"""
-    click.echo("Restoring files...")
-
-@git.command()
-def rm():
-    """Remove files from the working tree and from the index"""
-    click.echo("Removing files...")
-
-# --- Examine the History and State ---
-@git.command()
-def bisect():
-    """Use binary search to find the commit that introduced a bug"""
-    click.echo("Running bisect...")
-
-@git.command()
-def diff():
-    """Show changes between commits, commit, and working tree"""
-    click.echo("Showing differences...")
-
-@git.command()
-def grep():
-    """Print lines matching a pattern"""
-    click.echo("Searching for pattern...")
-
-@git.command()
-def log():
-    """Show commit logs"""
-    click.echo("Showing commit logs...")
-
-@git.command()
-def show():
-    """Show various types of objects"""
-    click.echo("Showing objects...")
-
-@git.command()
-def status():
-    """Show the working tree status"""
-    click.echo("Showing status...")
-
-# --- Grow, Mark and Tweak History ---
-@git.command()
-def branch():
-    """List, create, or delete branches"""
-    click.echo("Managing branches...")
-
-@git.command()
-def commit():
-    """Record changes to the repository"""
-    click.echo("Committing changes...")
-
-@git.command()
-def merge():
-    """Join two or more development histories together"""
-    click.echo("Merging branches...")
-
-@git.command()
-def rebase():
-    """Reapply commits on top of another base tip"""
-    click.echo("Rebasing...")
-
-@git.command()
-def reset():
-    """Reset current HEAD to the specified state"""
-    click.echo("Resetting HEAD...")
-
-@git.command()
-def switch():
-    """Switch branches"""
-    click.echo("Switching branches...")
-
-@git.command()
-def tag():
-    """Create, list, delete, or verify a tag object signed with GPG"""
-    click.echo("Managing tags...")
-
-# --- Collaborate with Remote ---
-@git.command()
-def fetch():
-    """Download objects and refs from another repository"""
-    click.echo("Fetching data...")
-
-@git.command()
-def pull():
-    """Fetch from and integrate with another repository or a local branch"""
-    click.echo("Pulling changes...")
-
-@git.command()
-def push():
-    """Update remote refs along with associated objects"""
-    click.echo("Pushing changes...")
-
-if __name__ == "__main__":
-    git()
